Move buttons to EyeTracking, fix hedgehogs showing
